Update the update script to avoid merges
With the previous update script, whenever someone retracted a commit the
local repository of the website server would divert its history and
start doing merge commits everytime we pulled.

We solve this by doing a 'fetch' and 'checkout origin/master' instead.
That enforces that the local repo of the website server is a mirror of
the GitHub repo.